LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/unloading-documents/001-expected.txt

main frame - has 1 onunload handler(s)
frame "<!--frame1-->" - has 1 onunload handler(s)
CONSOLE MESSAGE: line 21: TypeError: do_test is not a function. (In 'do_test()', 'do_test' is undefined)

FAIL document.open in unload assert_equals: expected "0123456789" but got "012389"
